comfyUI_ctr_reciever/requirements.txt
# Requirements for ComfyUI Controller Receiver Nodes
# Install these in your ComfyUI Python environment:
#   pip install websocket-client

websocket-client>=1.6.0
